Adding and changing files for GUI (#60)
* Adding and changing files for GUI

Adding the ResultFrame file, and making any necessary changes for it to work.

* Create temp file for directory

* Add Grpc for coordinator

* Appeasing checkstyle

* Remove unused implementation, fix imports

Originally, I had this using ActionListener, but decided against it and forgot to remove it. Also fixed imports for checkstyle

* Removing getFirst()

This for some reason isn't compiling here, but working perfectly fine when I run it locally. I'll try changing it to get(0) to see if that helps but I'm really not sure why there's a problem here.